fix: After deleting an update, the next new update will not show
When operating normally, the JSONL database always terminates with a
newline. When deleting an update, because the file was completely
re-written (truncated) with a `updates |> map(json.serialize) |>
join('\n')` call, the terminating newline disappeared.

This caused a bug where after deleting an update the next update would
simply be appended to the last update, silently failing. This fixes that
bug.